DEV Add a utils module to separate utilities
serpentTools.utils now exposes some more developer-focused
utilities and clears up a lot of the junk, i.e. all the
specific variables for the plot decorator.
Inside this module we have all the utilies from before, but spread
into a few different files:

    docstrings: functions for modifying docstrings, e.g. our
        magicPlotDocDecorator
    plot: functions for altering plots, e.g. our formatPlot function

__init__ file contains our str2Vec and convertVariableName functions

Minimal tweaks to some of our files. Basically, unless the
cartMeshPlot or generic plot function are called, all imports from
serpentTools.utils are valid.
